refactor : 유틸함수 date.ts 로 위치 변경
