+++ src/shared/utils/date.ts
@@ -1,1 +1,9 @@
 export const minutesToMs = (minutes: number) => minutes * 60 * 1000;
+
+export const formatDate = (dateString: string) => {
+  const date = new Date(dateString);
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${year} - ${month} - ${day}`;
+};
